Tweak debug_redact behavior (#460)
* Tweak debug_redact behavior

In sansshell we use redact to avoid logging sensitive info primarily for avoiding leaking secrets, not to provide privacy. I've switched our redaction from `--REDACTED--` to `REDACTED-<sha256hash>` so that it's possible to figure out if a field matches a known value while not being possible to learn the value purely from reading the field. The empty value is left unredacted because it should never be secret. I've also added support for bytes in addition to strings.

Now that we have a more generous redaction, I've added it to a couple more fields that seem like they might accidentally contain secrets.

* Add comment and more tests