Remove global variable storing the gRPC server object
The global variable 'srv' has been used in one test case to stop a
server which has been spawned in the test. As we are preparing to
expose a Unix socket, the global variable is problematic because there
are going to be _two_ distinct servers: one for the TCP socket and one
for the Unix socket.

We modify the server code and the test case to use a hook function
which gets the server object as an argument, rather than having to
access the server object directly via the global variable.

This allows us to get rid of the mutex in the `server` module, too.