Expose Unix Domain Socket from server for local access (#494)
* Remove global variable storing the gRPC server object

The global variable 'srv' has been used in one test case to stop a
server which has been spawned in the test. As we are preparing to
expose a Unix socket, the global variable is problematic because there
are going to be _two_ distinct servers: one for the TCP socket and one
for the Unix socket.

We modify the server code and the test case to use a hook function
which gets the server object as an argument, rather than having to
access the server object directly via the global variable.

This allows us to get rid of the mutex in the `server` module, too.

* Add server option to expose Unix socket for local gRPC access

Apart from providing access to a node from the outside, Sansshell
could conceivably be viewed as an interface between services running
on the node and the node's operating system. For example, instead of
adding a sudoers entry to allow some operation which would normally
require root privileges, we could expose the operation as a Sansshell
gRPC method, and call it from the inside: from some process which runs
on the same node as the Sansshell server. This way, we would gain
fine-grained authorization and auditability.

To make this approach possible, we modify the Sansshell server to
optionally expose a Unix domain socket which offers the same gRPC
methods as the main TCP server. The server which handles the socket
does not require TLS authentication, because obtaining the proper
certificates may be difficult for local clients. This should be OK,
considering how MITM attacks using the socket would only be possible
if the node has already been compromised to some extent.

Unix permissions/ACLs on the socket file can be used as a security
measure. The server makes this possible by exposing a hook function
which can configure the socket's ownership and permissions once the
socket is created. The hook mechanism may be used in the future to
specify socket permissions and ownership on the server command line.

In the rego authentication rules, clients using the Unix socket can be
identified by having 'input.peer.net.network = "unix"'.