Pass peer credentials from Unix domain socket to authorization engine
In commit 39af558b10, we have added the possibility of exposing a Unix
domain socket to the Sansshell server. In this commit, we enhance the
authentication and authorization possibilities around this new method
of communication.

On Linux systems, we can get information about the process which has
initiated a connection over a Unix socket by means of the `getsockopt`
call with the `SO_PEERCRED` option. This way, we get the UID, GID and
PID of the calling process. We pass this information into the input
structure of the OPA rules engine, so that rules can be written to
only allow certain local users to access a particular Sansshell gRPC
method, for example.

An equivalent mechanism via the `getsockopt` option `LOCAL_PEERCRED``
is available on Darwin (macOS) systems, so we include Unix credentials
in the auth input there as well for the sake of completeness. This
also aids testability during development.

In addition to the numeric UID and GID values in the credentials of a
peer talking to the Sansshell server over a Unix socket, we provide
the human-readable user and group names. This will enable writing more
reader-friendly OPA rules based on the Unix credentials.