Allow passing extra file descriptors through util.RunCommand
Some gRPC methods in SansShell may require passing an open file
descriptor through to an external command executed by SansShell. This
is useful for passing secrets in some cases: when the receiving
program accepts a file descriptor, we can create a descriptor for an
ephemeral in-memory file and pass the descriptor down to the
program. This approach allows us to avoid creating a file on disk with
the secret or passing the secret through command-line parameters.

An example usage of this might be mounting an `ecryptfs` volume: the
`mount.ecryptfs` helper accepts the `-o key=passphrase_passwd_fd=<fd>`
mount option, and using it looks safer than the alternatives of
passing the secret through a named file or through command-line
parameters.